+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/data/Follower.kt
@@ -1,0 +1,4 @@
+package com.bounswe2020group3.paperlayer.data
+
+class Follower {
+}
