--- android/app/src/main/java/com/bounswe2020group3/paperlayer/data/follow/FollowCreate.kt
@@ -1,10 +1,0 @@
-package com.bounswe2020group3.paperlayer.data.follow
-
-import com.squareup.moshi.Json
-
-data class FollowCreate (
-    @field:Json(name = "from_user")
-    var fromUser: Int,
-    @field:Json(name = "to_user")
-    var toUser: Int,
-) {}
