--- android/app/src/main/java/com/bounswe2020group3/paperlayer/data/follow/FollowRequest.kt
@@ -1,13 +1,0 @@
-package com.bounswe2020group3.paperlayer.data.follow
-
-import com.bounswe2020group3.paperlayer.data.user.User
-import com.squareup.moshi.Json
-
-data class FollowRequest (
-    @field:Json(name = "req_from_user")
-    var requestFromUser: User,
-    @field:Json(name = "req_to_user")
-    var requestToUser: User,
-    @field:Json(name = "created")
-    var created: String,
-) {}
