--- android/app/src/main/java/com/bounswe2020group3/paperlayer/data/follow/FollowRequestCreate.kt
@@ -1,10 +1,0 @@
-package com.bounswe2020group3.paperlayer.data.follow
-
-import com.squareup.moshi.Json
-
-data class FollowRequestCreate(
-    @field:Json(name = "req_from_user")
-    var requestFromUser: Int,
-    @field:Json(name = "req_to_user")
-    var requestToUser: Int,
-) {}
