+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/ProfileFragment.kt
@@ -6,13 +6,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
 import com.bounswe2020group3.paperlayer.MainActivity
 import com.bounswe2020group3.paperlayer.R
-import com.bounswe2020group3.paperlayer.data.follow.FollowType
-import com.bounswe2020group3.paperlayer.data.user.User
+import com.bounswe2020group3.paperlayer.profile.data.User
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.fragment_profile.*
 import javax.inject.Inject
@@ -44,15 +42,12 @@ class ProfileFragment : Fragment(), ProfileContract.View {
             Navigation.findNavController(view).navigate(R.id.navigateToProfileEditFromProfile)
         }
 
-        val followerBundle = bundleOf("followType" to FollowType.FOLLOWER)
-        val followingBundle = bundleOf("followType" to FollowType.FOLLOWING)
-
         linearLayoutFollowers.setOnClickListener{
-            Navigation.findNavController(view).navigate(R.id.navigateToFollowListFromProfile, followerBundle)
+            Navigation.findNavController(view).navigate(R.id.navigateToUserListFromProfile)
         }
 
         linearLayoutFollowings.setOnClickListener{
-            Navigation.findNavController(view).navigate(R.id.navigateToFollowListFromProfile, followingBundle)
+            Navigation.findNavController(view).navigate(R.id.navigateToUserListFromProfile)
         }
     }
