+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/data/data/UserCredentials.kt
@@ -1,4 +1,4 @@
-package com.bounswe2020group3.paperlayer.data.user
+package com.bounswe2020group3.paperlayer.profile.data.data
 
 import com.squareup.moshi.Json
 data class UserCredentials (
