+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/user/UserContract.kt
@@ -1,13 +1,12 @@
 package com.bounswe2020group3.paperlayer.profile.user
 
 import com.bounswe2020group3.paperlayer.mvp.Mvp
-import com.bounswe2020group3.paperlayer.data.user.User
+import com.bounswe2020group3.paperlayer.profile.data.Profile
+import com.bounswe2020group3.paperlayer.profile.data.User
 
 interface UserContract {
     interface Presenter : Mvp.Presenter<View> {
         fun loadUser(userId: Int)
-        fun sendFollow(toUserId: Int)
-        fun sendFollowRequest(toUserId: Int)
     }
 
     interface View : Mvp.View {
@@ -17,6 +16,5 @@ interface UserContract {
         fun showInfoToast(message: String = "Info")
         fun showErrorToast(message: String = "Error")
         fun updateProfileUI(user: User)
-        fun loadUser()
     }
 }
