+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/user/UserFragment.kt
@@ -7,23 +7,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.core.os.bundleOf
 import androidx.navigation.Navigation
 import com.bounswe2020group3.paperlayer.MainActivity
 import com.bounswe2020group3.paperlayer.R
-import com.bounswe2020group3.paperlayer.data.follow.FollowType
-import com.bounswe2020group3.paperlayer.data.user.User
+import com.bounswe2020group3.paperlayer.profile.data.User
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.fragment_user.*
-import kotlinx.android.synthetic.main.fragment_user.imageViewProfileAvatar
-import kotlinx.android.synthetic.main.fragment_user.textViewBio
-import kotlinx.android.synthetic.main.fragment_user.textViewBirthday
-import kotlinx.android.synthetic.main.fragment_user.textViewEmail
-import kotlinx.android.synthetic.main.fragment_user.textViewExpertise
-import kotlinx.android.synthetic.main.fragment_user.textViewFullName
-import kotlinx.android.synthetic.main.fragment_user.textViewGender
-import kotlinx.android.synthetic.main.fragment_user.textViewInterests
-
 import javax.inject.Inject
 
 private const val ARG_USER_ID = "userID"
@@ -32,8 +21,7 @@ class UserFragment : Fragment(), UserContract.View {
 
     @Inject lateinit var presenter: UserContract.Presenter
 
-    private var userID: Int? = -1
-    private var isFollowRequest = true
+    private var userID: Int? = null
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
@@ -57,35 +45,14 @@ class UserFragment : Fragment(), UserContract.View {
         return view
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        buttonUserFollow.setOnClickListener {
-            if(this.userID != null && this.userID!! > 0) {
-                if(this.isFollowRequest) {
-                    presenter.sendFollowRequest(userID!!)
-                } else {
-                    presenter.sendFollow(userID!!)
-                }
-            }
-        }
-
-        val followerBundle = bundleOf("followType" to FollowType.FOLLOWER, "userID" to userID)
-
-        val followingBundle = bundleOf("followType" to FollowType.FOLLOWING, "userID" to userID)
-
-        layoutUserFollowers.setOnClickListener{
-            Navigation.findNavController(view).navigate(R.id.navigateToFollowListFromUser, followerBundle)
-        }
-
-        layoutUserFollowings.setOnClickListener{
-            Navigation.findNavController(view).navigate(R.id.navigateToFollowListFromUser, followingBundle)
-        }
-    }
-
     override fun onResume() {
         super.onResume()
-        this.loadUser()
+        if(userID != null) {
+            presenter.loadUser(userID!!.toInt())
+        } else {
+            showErrorToast("User id is invalid. Please try again.")
+            navigateBack()
+        }
     }
 
     companion object {
@@ -124,74 +91,18 @@ class UserFragment : Fragment(), UserContract.View {
     override fun updateProfileUI(user: User) {
         val profile = user.profile[0]
         val fullName = "${profile.name} ${profile.lastName}"
+
+        textViewEmail.text = user.email
         textViewFullName.text = fullName
-
-        if(profile.is_public!! || user.isFollowing) {
-            // Public or following profile
-            // Private Profile
-            this.isFollowRequest = false
-
-            // Layouts
-            layoutUserStatsWrapper.visibility = View.VISIBLE
-            layoutUserInformationWrapper.visibility = View.VISIBLE
-            layoutUserPrivateWrapper.visibility = View.GONE
-
-            // Bio
-            if(profile.bio == null || profile.bio == "") {
-                layoutUserBioWrapper.visibility = View.GONE
-            } else {
-                layoutUserBioWrapper.visibility = View.VISIBLE
-                textViewBio.text = profile.bio
-            }
-
-            // Text Views
-            textViewEmail.text = user.email
-            textViewBirthday.text = profile.birthday.toString()
-            textViewGender.text = profile.gender
-            textViewInterests.text = profile.interests
-            textViewExpertise.text = profile.expertise
-
-            // Buttons
-            buttonUserRequestSent.visibility = View.GONE
-            if(user.isFollowing) {
-                buttonUserFollow.visibility = View.GONE
-                buttonUserUnfollow.visibility = View.VISIBLE
-            } else {
-                buttonUserFollow.visibility = View.VISIBLE
-                buttonUserUnfollow.visibility = View.GONE
-            }
-        } else {
-            // Private Profile
-            this.isFollowRequest = true
-
-            layoutUserStatsWrapper.visibility = View.GONE
-            layoutUserBioWrapper.visibility = View.GONE
-            layoutUserInformationWrapper.visibility = View.GONE
-            layoutUserPrivateWrapper.visibility = View.VISIBLE
-
-            // Buttons
-            buttonUserUnfollow.visibility = View.GONE
-            if(user.isFollowRequestSent) {
-                buttonUserRequestSent.visibility = View.VISIBLE
-                buttonUserFollow.visibility = View.GONE
-            } else {
-                buttonUserRequestSent.visibility = View.GONE
-                buttonUserFollow.visibility = View.VISIBLE
-            }
-        }
+        textViewBio.text = profile.bio
+        textViewBirthday.text = profile.birthday.toString()
+        textViewGender.text = profile.gender
+        textViewInterests.text = profile.interests
+        textViewExpertise.text = profile.expertise
 
         val imageUrl = profile.profile_picture
         if(imageUrl != null && imageUrl.contains("http")) {
             Picasso.get().load(imageUrl).into(imageViewProfileAvatar)
-        }
-    }
-
-    override fun loadUser() {
-        if(userID != null && userID!! > 0) {
-            presenter.loadUser(userID!!)
-        } else {
-            showErrorToast("User id is invalid. Please try again.")
-            navigateBack()
         }
     }
