+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/util/Session.kt
@@ -1,6 +1,6 @@
 package com.bounswe2020group3.paperlayer.util
 
-import com.bounswe2020group3.paperlayer.data.user.AuthToken
+import com.bounswe2020group3.paperlayer.profile.data.data.AuthToken
 import io.reactivex.subjects.BehaviorSubject
 
 interface Session {
