+++ android/app/src/main/res/navigation/main_navigation.xml
@@ -58,9 +58,6 @@
         <action
             android:id="@+id/navigateToUserListFromProfile"
             app:destination="@id/userListFragment" />
-        <action
-            android:id="@+id/navigateToFollowListFromProfile"
-            app:destination="@id/followListFragment" />
     </fragment>
 
     <fragment
@@ -115,11 +112,7 @@
         android:id="@+id/searchFragment"
         android:name="com.bounswe2020group3.paperlayer.search.SearchFragment"
         android:label="fragment_search"
-        tools:layout="@layout/fragment_search" >
-        <action
-            android:id="@+id/navigateToUserFromSearch"
-            app:destination="@id/userFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/userListFragment"
         android:name="com.bounswe2020group3.paperlayer.profile.list.UserListFragment"
@@ -136,9 +129,6 @@
         <argument
             android:name="userID"
             app:argType="integer" />
-        <action
-            android:id="@+id/navigateToFollowListFromUser"
-            app:destination="@id/followListFragment" />
     </fragment>
     <fragment
         android:id="@+id/inviteFragment"
@@ -178,16 +168,5 @@
         <action
             android:id="@+id/navigateToEventsFromProjectUpdates"
             app:destination="@id/eventFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/followListFragment"
-        android:name="com.bounswe2020group3.paperlayer.profile.follow.FollowListFragment"
-        android:label="FollowListFragment" >
-        <action
-            android:id="@+id/navigateToUserFromFollowList"
-            app:destination="@id/userFragment" />
-        <action
-            android:id="@+id/navigateToProfileFromFollowList"
-            app:destination="@id/profileFragment" />
     </fragment>
 </navigation>
