+++ android/app/src/main/res/values/colors.xml
@@ -10,6 +10,5 @@
     <color name="colorBlack">#000000</color>
     <color name="colorBlackLowAlpha">#34000000</color>
     <color name="colorGray">#E5E5E5</color>
-    <color name="colorDarkGray">#252525</color>
     <color name="colorTransparent">#00FFFFFF</color>
 </resources>
